requirements.txt
feedparser
requests
pyyaml
trafilatura
lxml_html_clean
urllib3<2
beautifulsoup4
